Cleanup

+++ src/main/scala/ai/search/Heuristic.scala
@@ -1,5 +1,5 @@
 package ai.search
 
 trait Heuristic[A] {
-  def h(state: A): Float
+  def h(state: A): Double
 }
